requirements.txt
black
channels
Django
djangorestframework
djangorestframework-simplejwt
pytest
python-dotenv